Move common code to common folder
