+++ default/python/AI/ColonisationAI.py
@@ -13,7 +13,7 @@ import MilitaryAI
 from turn_state import state
 from EnumsAI import MissionType, FocusType, EmpireProductionTypes, ShipRoleType, PriorityType
 from freeorion_tools import dict_from_map, tech_is_complete, get_ai_tag_grade, cache_by_turn
-from freeorion_debug import Timer
+from common.timers import Timer
 
 colonization_timer = Timer('getColonyFleets()')
