--- default/python/AI/freeorion_debug/__init__.py
@@ -1,9 +1,0 @@
-"""
-This module will store all tool for debug and testing.
-"""
-
-import freeorion_debug.extend_free_orion_AI_interface
-from freeorion_debug.interactive_shell import handle_debug_chat
-from freeorion_debug.listeners import listener
-from freeorion_debug.timers import Timer
-import freeorion_debug.handlers
