--- default/python/AI/freeorion_debug/handlers/inspect_interface_config.ini
@@ -1,2 +1,0 @@
-[main]
-handlers=freeorion_debug/handlers/inspect_freeOrionAIInterface.py
