+++ default/python/common/handlers/inspect_freeOrionAIInterface.py
@@ -2,7 +2,7 @@
 import PlanetUtilsAI
 import freeOrionAIInterface as fo
 
-from freeorion_debug.ide_tools import inspect
+from common.ide_tools import inspect
 
 
 def inspect_ai_interface():
@@ -63,6 +63,6 @@ def inspect_ai_interface():
     )
     exit(1)  # exit game to main menu no need to play anymore.
 
-from freeorion_debug.listeners import register_pre_handler
+from common.listeners import register_pre_handler
 
 register_pre_handler('generateOrders', inspect_ai_interface)
