+++ default/python/common/handlers/inspect_interface_config.ini
@@ -1,0 +1,2 @@
+[main]
+handlers=common/handlers/inspect_freeOrionAIInterface.py
