+++ default/python/common/ide_tools/generate_stub.py
