+++ default/python/common/ide_tools/result/freeorion.py
